Develop

+++ BSP/bsp.h
@@ -142,7 +142,7 @@ typedef struct
         struct Spi2
         {
             void (*WriteData)(uint8_t data);
-            void (*WriteNData)(uint8_t *data,uint8_t len);
+            void (*WriteNData)(uint8_t *data,uint16_t len);
             uint8_t (*WriteReadData)(uint8_t data);
         }Spi2;
